- Finished writing and debugging the AES128 class 	- Decryption methods still have to be figured out - Fixed cvtStrToState 	- It will now assign elements column-wise, as opposed to row-wise - Did more cleaning up of comments and function definitions - Added TODOs

+++ encrypt/AES.h
@@ -11,24 +11,27 @@
 class AES
 {
 protected:
-    static const int numRows {4};
-    static const int Nb {4};
+    static const size_t numRows {4};
+    static const size_t Nb {4};
     unsigned char state[numRows][Nb];
 
-    // Functions used in the Rijndael cipher
+    // AES ecnryption functions
     void byteSub();
     void shiftRow();
     void mixColumns();
 
+    // Helper functions
     std::string cvtStateToStr();
     void cvtStrToState(std::string plain);
 
+    // Pure virtual functions
     virtual void addRoundKey(unsigned char* key, int round) = 0;
     virtual void cvtStrToKey(std::string in) = 0;
     virtual void finalRound(int round) = 0;
     virtual void round(int round) = 0;
 
 public:
+    // Main encryption and decryption functions
     virtual std::string cipher(std::string plainText, std::string keyIn) = 0;
 
 };
